Data EDA and Processing
We condensed 2.5 GBs of New York Crime data into a condensed 811 KB file which incudes crime/data ratings of 14000 nodes on the map of Manhattan, We removed all crimes unrelated to the safety of valuable truck or a car carrying a person of significance, we also removed all data before 2016(40% of entries) as well as all data related to age/sex/race of the victim as well as the suspect, description/time/case status/process/case number/jurisdiction code/nearby parks as well as other non required column. Fot the remaining type of crimes we assigned a rating between 1-10.0. These crimes had a latitude longitude coordinate which we termed as nodes, these nodes are at intersection of roads and middle of roads. We take a cumulative sum of the ratings of all the crime happening at each of these nodes, then took a log of the ratings and normalized that data to a rating from 1-10 (where 1 signifies a safe road, and 10 is an extremely dangerous road. We wanted to use these normalized nodes as speed limiters as in where a car is hypothetically allowed to go at 100/(node rating) kmph on the particular node, so on a safe road (1 rating) it can go at 100 kmph and just and 10 kphm on a dangerous (10 rating) road. After this we would apply a Shortest Time algorithm to find the fastest( safest) route for the vehicle.  We would provide 3 possible routes, The Shortest distance route, the Safest route and a compromise between the two but were unable to link our dataset with any available dataset.